Split out MessagePrompts
