+++ packages/mongodb-chatbot-ui/src/Message.tsx
@@ -10,15 +10,10 @@ import {
   MessageSourceType,
 } from "@lg-chat/message";
 import {
-  MessagePrompts as LGMessagePrompts,
-  MessagePrompt,
-} from "@lg-chat/message-prompts";
-import {
   type MessageRatingProps,
   type MessageRatingValue,
 } from "@lg-chat/message-rating";
-import { Fragment, lazy, Suspense, useRef, useState } from "react";
-import { CSSTransition } from "react-transition-group";
+import { Fragment, lazy, Suspense, useState } from "react";
 import { useUser, User } from "./useUser";
 import { MessageData } from "./services/conversations";
 import { Conversation } from "./useConversation";
@@ -30,6 +25,10 @@ import { type RatingCommentStatus } from "./MessageRating";
 
 const MessageRatingWithFeedbackComment = lazy(async () => ({
   default: (await import("./MessageRating")).MessageRatingWithFeedbackComment,
+}));
+
+const MessagePrompts = lazy(async () => ({
+  default: (await import("./MessagePrompts")).MessagePrompts,
 }));
 
 const TRANSITION_DURATION_MS = 300;
@@ -309,78 +308,15 @@ export const Message = ({
           ) : null}
         </Suspense>
       </LGMessage>
-      {showSuggestedPrompts && (
-        <MessagePrompts
-          prompts={suggestedPrompts}
-          onPromptClick={(prompt) => onSuggestedPromptClick?.(prompt)}
-          canSubmit={canSubmitSuggestedPrompt}
-        />
-      )}
+      <Suspense fallback={null}>
+        {showSuggestedPrompts && (
+          <MessagePrompts
+            prompts={suggestedPrompts}
+            onPromptClick={(prompt) => onSuggestedPromptClick?.(prompt)}
+            canSubmit={canSubmitSuggestedPrompt}
+          />
+        )}
+      </Suspense>
     </Fragment>
-  );
-};
-
-export type MessagePromptsProps = {
-  prompts: string[];
-  onPromptClick: (prompt: string) => void;
-  canSubmit: (prompt: string) => boolean;
-};
-
-export const MessagePrompts = ({
-  prompts,
-  onPromptClick,
-  canSubmit,
-}: MessagePromptsProps) => {
-  const [selectedPromptIndex, setSelectedPromptIndex] = useState<
-    number | undefined
-  >(undefined);
-  const nodeRef = useRef(null);
-
-  // This ref is used to prevent the user from clicking a suggested
-  // prompt multiple times. We use a ref instead of state to ensure that
-  // the prompt is only selected and sent to the server once regardless
-  // of where we are in the React render cycle.
-  const suggestedPromptClickedRef = useRef(false);
-  const onPromptSelected = (prompt: string, idx: number) => {
-    // Don't do anything if the prompt is not selectable.
-    if (!canSubmit(prompt)) {
-      return;
-    }
-    // Check the ref to prevent the prompt from being clicked multiple
-    // times. This might happen if the user clicks the prompt again
-    // while the list of prompts is animating out.
-    if (suggestedPromptClickedRef.current) {
-      return;
-    }
-    suggestedPromptClickedRef.current = true;
-    setSelectedPromptIndex(idx);
-    // Wait for the prompts to fully animate out before calling the
-    // click handler.
-    setTimeout(() => {
-      onPromptClick(prompt);
-    }, TRANSITION_DURATION_MS);
-  };
-
-  return (
-    <CSSTransition
-      in={selectedPromptIndex === undefined}
-      timeout={TRANSITION_DURATION_MS}
-      nodeRef={nodeRef}
-      classNames={styles.message_prompts}
-    >
-      <div className={styles.message_prompts} ref={nodeRef}>
-        <LGMessagePrompts label="Suggested Prompts">
-          {prompts.map((prompt, idx) => (
-            <MessagePrompt
-              key={prompt}
-              onClick={() => onPromptSelected(prompt, idx)}
-              selected={idx === selectedPromptIndex}
-            >
-              {prompt}
-            </MessagePrompt>
-          ))}
-        </LGMessagePrompts>
-      </div>
-    </CSSTransition>
   );
 };
